--- a/package.json
+++ b/package.json
@@ -40,6 +40,7 @@
     "@types/long": "^4.0.0",
     "@types/mocha": "^5.2.5",
     "@types/node": "^10.7.0",
+    "benchmark": "^2.1.4",
     "chai": "^4.2.0",
     "codecov": "^3.0.4",
     "mocha": "^5.2.0",