--- a/package.json
+++ b/package.json
@@ -53,6 +53,6 @@
   "dependencies": {
     "extensible-error": "^1.0.2",
     "long": "^4.0.0",
-    "oer-utils": "^3.2.0"
+    "oer-utils": "^4.0.0"
   }
 }