--- a/src/errors.ts
+++ b/src/errors.ts
@@ -102,7 +102,7 @@ export class AmountTooLargeError extends BaseError {
     super(message)
     this.ilpErrorCode = codes.F08_AMOUNT_TOO_LARGE
 
-    const writer = new Writer()
+    const writer = new Writer(8 + 8)
     const receivedAmountLong = Long.fromString(opts.receivedAmount, true)
     const maximumAmountLong = Long.fromString(opts.maximumAmount, true)
     writer.writeUInt32(receivedAmountLong.getHighBitsUnsigned())