Depends on https://github.com/interledgerjs/oer-utils/pull/21.

Benchmarks: (before and after):

```
serializeIlpPrepare      v0 x 224,294 ops/sec ±0.84% (91 runs sampled)
serializeIlpPrepare      v1 x 418,535 ops/sec ±0.88% (91 runs sampled)
deserializeIlpPrepare    v0 x 236,698 ops/sec ±1.28% (91 runs sampled)
deserializeIlpPrepare    v1 x 322,443 ops/sec ±0.45% (91 runs sampled)
serializeIlpFulfill      v0 x 537,387 ops/sec ±0.86% (90 runs sampled)
serializeIlpFulfill      v1 x 1,054,131 ops/sec ±1.20% (88 runs sampled)
deserializeIlpFulfill    v0 x 1,079,358 ops/sec ±0.68% (94 runs sampled)
deserializeIlpFulfill    v1 x 4,630,731 ops/sec ±1.60% (89 runs sampled)
serializeIlpReject       v0 x 260,279 ops/sec ±0.82% (88 runs sampled)
serializeIlpReject       v1 x 457,782 ops/sec ±0.98% (91 runs sampled)
deserializeIlpReject     v0 x 478,236 ops/sec ±0.69% (97 runs sampled)
deserializeIlpReject     v1 x 1,333,481 ops/sec ±1.00% (89 runs sampled)
```